Moved actor input / actor items taking to another manager.
