+++ src/engine/core/managers/actor/ActorInputManager.ts
@@ -3,9 +3,20 @@ import { game, get_hud, level } from "xray16";
 import { closeLoadMarker, closeSaveMarker, openLoadMarker, openSaveMarker, registry } from "@/engine/core/database";
 import { AbstractManager } from "@/engine/core/managers/base";
 import { EGameEvent, EventsManager } from "@/engine/core/managers/events";
+import { SurgeManager } from "@/engine/core/managers/surge";
+import { surgeConfig } from "@/engine/core/managers/surge/SurgeConfig";
+import { killAllSurgeUnhidden } from "@/engine/core/managers/surge/utils";
+import { WeatherManager } from "@/engine/core/managers/weather";
+import { executeConsoleCommand, getConsoleFloatCommand } from "@/engine/core/utils/game";
 import { readTimeFromPacket, writeTimeToPacket } from "@/engine/core/utils/game/game_time";
 import { LuaLogger } from "@/engine/core/utils/logging";
+import { disableInfo, giveInfo } from "@/engine/core/utils/object";
 import { isActorInNoWeaponZone } from "@/engine/core/utils/object/object_zone";
+import { animations, postProcessors } from "@/engine/lib/constants/animation";
+import { consoleCommands } from "@/engine/lib/constants/console_commands";
+import { infoPortions } from "@/engine/lib/constants/info_portions";
+import { TInventoryItem } from "@/engine/lib/constants/items";
+import { drugs } from "@/engine/lib/constants/items/drugs";
 import { misc } from "@/engine/lib/constants/items/misc";
 import {
   ClientObject,
@@ -14,9 +25,11 @@ import {
   NetPacket,
   NetProcessor,
   Optional,
+  ServerObject,
   TDuration,
   Time,
   TIndex,
+  TRate,
 } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename);
@@ -41,7 +54,8 @@ export class ActorInputManager extends AbstractManager {
 
     eventsManager.registerCallback(EGameEvent.ACTOR_UPDATE, this.onUpdate, this);
     eventsManager.registerCallback(EGameEvent.ACTOR_FIRST_UPDATE, this.onFirstUpdate, this);
-    eventsManager.registerCallback(EGameEvent.ACTOR_GO_ONLINE, this.onNetworkSpawn, this);
+    eventsManager.registerCallback(EGameEvent.ACTOR_GO_ONLINE, this.onActorGoOnline, this);
+    eventsManager.registerCallback(EGameEvent.ACTOR_USE_ITEM, this.onActorUseItem, this);
   }
 
   public override destroy(): void {
@@ -49,7 +63,8 @@ export class ActorInputManager extends AbstractManager {
 
     eventsManager.unregisterCallback(EGameEvent.ACTOR_UPDATE, this.onUpdate);
     eventsManager.unregisterCallback(EGameEvent.ACTOR_FIRST_UPDATE, this.onFirstUpdate);
-    eventsManager.unregisterCallback(EGameEvent.ACTOR_GO_ONLINE, this.onNetworkSpawn);
+    eventsManager.unregisterCallback(EGameEvent.ACTOR_GO_ONLINE, this.onActorGoOnline);
+    eventsManager.unregisterCallback(EGameEvent.ACTOR_USE_ITEM, this.onActorUseItem);
   }
 
   public override save(packet: NetPacket): void {
@@ -236,6 +251,24 @@ export class ActorInputManager extends AbstractManager {
   }
 
   /**
+   * todo: Description.
+   */
+  public processAnabioticItemUsage(): void {
+    ActorInputManager.getInstance().disableGameUiOnly();
+
+    level.add_cam_effector(animations.camera_effects_surge_02, 10, false, "engine.on_anabiotic_sleep");
+    level.add_pp_effector(postProcessors.surge_fade, 11, false);
+
+    giveInfo(infoPortions.anabiotic_in_process);
+
+    registry.sounds.musicVolume = getConsoleFloatCommand(consoleCommands.snd_volume_music);
+    registry.sounds.effectsVolume = getConsoleFloatCommand(consoleCommands.snd_volume_eff);
+
+    executeConsoleCommand(consoleCommands.snd_volume_music, 0);
+    executeConsoleCommand(consoleCommands.snd_volume_eff, 0);
+  }
+
+  /**
    * Handle generic update from actor input perspective.
    */
   public onUpdate(delta: TDuration): void {
@@ -290,9 +323,88 @@ export class ActorInputManager extends AbstractManager {
   /**
    * Handle actor network spawn.
    */
-  public onNetworkSpawn(): void {
+  public onActorGoOnline(): void {
     if (this.disableInputAt === null) {
       level.enable_input();
     }
+  }
+
+  /**
+   * Handle actor item use.
+   * Mainly to intercept and properly handle anabiotic.
+   */
+  public onActorUseItem(object: Optional<ClientObject>): void {
+    if (object === null) {
+      return;
+    }
+
+    const serverObject: Optional<ServerObject> = registry.simulator.object(object.id());
+    const serverItemSection: Optional<TInventoryItem> = serverObject?.section_name() as Optional<TInventoryItem>;
+
+    if (serverItemSection === drugs.drug_anabiotic) {
+      logger.info("On actor anabiotic use:", object.name());
+      this.processAnabioticItemUsage();
+    }
+  }
+
+  /**
+   * todo: Description.
+   */
+  public onAnabioticSleep(): void {
+    level.add_cam_effector(animations.camera_effects_surge_01, 10, false, "engine.on_anabiotic_wake_up");
+
+    const random: number = math.random(35, 45);
+    const surgeManager: SurgeManager = SurgeManager.getInstance();
+
+    if (surgeConfig.IS_STARTED) {
+      const timeFactor: TRate = level.get_time_factor();
+      const timeDiffInSeconds: TDuration = math.ceil(
+        game.get_game_time().diffSec(surgeManager.initializedAt) / timeFactor
+      );
+
+      if (random > ((surgeConfig.DURATION - timeDiffInSeconds) * timeFactor) / 60) {
+        surgeConfig.IS_TIME_FORWARDED = true;
+        surgeManager.isUiDisabled = true;
+        killAllSurgeUnhidden();
+        surgeManager.endSurge();
+      }
+    }
+
+    level.change_game_time(0, 0, random);
+    WeatherManager.getInstance().forceWeatherChange();
+  }
+
+  /**
+   * todo: Description.
+   */
+  public onAnabioticWakeUp(): void {
+    ActorInputManager.getInstance().enableGameUi();
+
+    executeConsoleCommand(consoleCommands.snd_volume_music, registry.sounds.musicVolume);
+    executeConsoleCommand(consoleCommands.snd_volume_eff, registry.sounds.effectsVolume);
+
+    registry.sounds.effectsVolume = 0;
+    registry.sounds.musicVolume = 0;
+
+    disableInfo(infoPortions.anabiotic_in_process);
+  }
+
+  /**
+   * todo: Description.
+   */
+  public onSurgeSurviveStart(): void {
+    level.add_cam_effector(
+      animations.camera_effects_surge_01,
+      surgeConfig.SLEEP_CAM_EFFECTOR_ID,
+      false,
+      "engine.surge_survive_end"
+    );
+  }
+
+  /**
+   * todo: Description.
+   */
+  public onSurgeSurviveEnd(): void {
+    this.enableGameUi();
   }
 }
